Can now spawn multiple enemies and have different wave speeds
In the enemyPathing script, I created a function called SetWaveConfig. This gets the waveConfig data from any script it is called in (I call it in enemyspawner.cs). I removed the moveSpeed variable and replaced it with calling GetMoveSpeed from waveConfig.cs

I also made some adjustments to the coroutine code in enemySpawner.cs (adding the newEnemy handle and using it to call the SetWaveConfig function)