Enemies can now take damage,
I created a DamageDealer script that contains data on the value of damage an object can deal. This is now attached to the laser.

I also created an enemy script that contains health information and code for collision. When colliding with an object that has a DamageDealer script(in this case, the laser), the enemy will get the damage value from object and subtract it from the enemy health variable.