Created enemy death explosion particle system.
Downloaded a 3-by-1 sprite sheet. Used it in a material that was added to a particle system object.

Edited the particle system's Texture Sheet Animation settings to randomy cycle between the different icons of the spritesheet.

Added randomization to particle size, hue, and duration.

Instantiated particle in Enemy.cs script as part of the enemyDies method.